moved folders to appropriate positions
