+++ src/core/domain/constants/types.ts
