+++ src/core/domain/data/function_query.ts
