+++ src/core/domain/models/interfaces/must_have_tenant.ts
