+++ src/core/domain/models/tenant.ts
