+++ src/core/domain/services/auth_service.ts
@@ -1,4 +1,8 @@
-import { UserDto, UserSignInInput, UserSignUpInput } from "../../ui/models/user_dto";
+import {
+    UserDto,
+    UserSignInInput,
+    UserSignUpInput
+} from "../../../ui/models/user_dto";
 
 export interface IAuthService {
     signUp(
