+++ src/core/domain/services/base_service.ts
