+++ src/core/domain/services/tenant_service.ts
@@ -1,5 +1,5 @@
-import { TenantDto } from "../../ui/models/tenant_dto";
-import { Tenant } from "../model/tenant";
+import { TenantDto } from "../../../ui/models/tenant_dto";
+import { Tenant } from "../models/tenant";
 import { IBaseService } from "./base_service";
 
 export interface ITenantService extends IBaseService<Tenant> {
