+++ src/core/domain/services/user_service.ts
@@ -1,0 +1,19 @@
+import { UserSignUpInput } from "../../../ui/models/user_dto";
+import { User } from "../models/user";
+
+export interface IUserService {
+    create(user: UserSignUpInput): Promise<User>;
+    get(query: { id?: string; emailOrUsername?: string }): Promise<User>;
+    getAll(): Promise<User[]>;
+    pagedGetAll({
+        searchStr,
+        skip,
+        limit
+    }: {
+        searchStr?: string;
+        skip?: number;
+        limit?: number;
+    }): Promise<{ totalCount: number; items: User[] }>;
+    update(user: Partial<User>): Promise<void>;
+    delete(id: string): Promise<boolean>;
+}
