+++ src/core/domain/utils/writable.ts
