+++ src/infrastructure/config/inversify.config.ts
@@ -1,12 +1,12 @@
 import { EventDispatcher } from "event-dispatch";
 import { ContainerModule } from "inversify";
 // Interfaces & Types
-import { TYPES } from "../../domain/constants/types";
+import { TYPES } from "../../core/domain/constants/types";
 
 // Controllers
 import "../../ui/api/controllers/auth_controller";
 import "../../ui/api/controllers/tenant_controller";
-import { IFunctionQuery } from "../../domain/data/function_query";
+import { IFunctionQuery } from "../../core/domain/data/function_query";
 import { FunctionQuery } from "../db/function_query";
 
 export const referenceDataIoCModule = new ContainerModule(bind => {
