+++ src/ui/api/middleware/auth_middleware.ts
@@ -2,9 +2,9 @@ import { Request } from "express";
 import httpStatus from "http-status-codes";
 import { Container } from "inversify";
 import jwt from "jsonwebtoken";
-import { TYPES } from "../../../domain/constants/types";
-import { ITenantRepository } from "../../../domain/data/repositories";
-import { UserRole } from "../../../domain/model/user";
+import { TYPES } from "../../../core/domain/constants/types";
+import { ITenantRepository } from "../../../core/domain/data/repositories";
+import { UserRole } from "../../../core/domain/models/user";
 import { config as env } from "../../../infrastructure/config";
 import { iocContainer } from "../../../infrastructure/config/ioc";
 import { TenantRepository } from "../../../infrastructure/db/repositories/tenant_repository";
