+++ src/ui/models/tenant_dto.ts
@@ -7,7 +7,7 @@ import {
 } from "class-validator";
 import { Expose } from "class-transformer";
 
-import { MAX_NAME_LENGTH } from "../../domain/model/user";
+import { MAX_NAME_LENGTH } from "../../core/domain/models/user";
 import { BaseCreateEntityDto, BaseEntityDto } from "./base_dto";
 
 export class CreateTenantInput extends BaseCreateEntityDto {
