+++ test/infrastructure/db/repository/tenant_repository.test.ts
@@ -1,6 +1,6 @@
 import { expect } from "chai";
-import { ITenantRepository } from "../../../../src/domain/data/repositories";
-import { Tenant } from "../../../../src/domain/model/tenant";
+import { ITenantRepository } from "../../../../src/core/domain/data/repositories";
+import { Tenant } from "../../../../src/core/domain/models/tenant";
 import { iocContainer } from "../../../../src/infrastructure/config/ioc";
 import { TenantRepository } from "../../../../src/infrastructure/db/repositories/tenant_repository";
 import { cleanupDb } from "../../../setup";
